Lattice site offset as integers

See merge request materials-modeling/icet!532

+++ src/LatticeSite.hpp
@@ -25,7 +25,7 @@ public:
     @param index site index
     @param unitcellOffset offset of site relative to unit cell at origin in units of lattice vectors
     */
-    LatticeSite(const size_t index, const Eigen::Vector3d unitcellOffset)
+    LatticeSite(const size_t index, const Eigen::Vector3i unitcellOffset)
     {
         _index = index;
         _unitcellOffset = unitcellOffset;
@@ -38,13 +38,13 @@ public:
     void setIndex(size_t index) { _index = index; }
 
     /// Return offset relative to unit cell at origin in units of lattice vectors.
-    Eigen::Vector3d unitcellOffset() const { return _unitcellOffset; }
+    Eigen::Vector3i unitcellOffset() const { return _unitcellOffset; }
 
     /// Set offset relative to unit cell at origin in units of lattice vectors.
-    void setUnitcellOffset(Eigen::Vector3d offset) { _unitcellOffset = offset; }
+    void setUnitcellOffset(Eigen::Vector3i offset) { _unitcellOffset = offset; }
 
     /// Add offset relative to unit cell at origin in units of lattice vectors.
-    void addUnitcellOffset(Eigen::Vector3d offset) { _unitcellOffset += offset; }
+    void addUnitcellOffset(Eigen::Vector3i offset) { _unitcellOffset += offset; }
 
     /// Smaller than operator.
     bool operator<(const LatticeSite &other) const
@@ -81,21 +81,21 @@ public:
     }
 
     /// Addition operator.
-    friend LatticeSite operator+(const LatticeSite &latticeSite, const Eigen::Vector3d &offset)
+    friend LatticeSite operator+(const LatticeSite &latticeSite, const Eigen::Vector3i &offset)
     {
         LatticeSite latnbr = LatticeSite(latticeSite.index(), latticeSite.unitcellOffset() + offset);
         return latnbr;
     }
 
     /// Substraction operator.
-    friend LatticeSite operator-(const LatticeSite &latticeSite, const Eigen::Vector3d &offset)
+    friend LatticeSite operator-(const LatticeSite &latticeSite, const Eigen::Vector3i &offset)
     {
         LatticeSite latnbr = LatticeSite(latticeSite.index(), latticeSite.unitcellOffset() - offset);
         return latnbr;
     }
 
     /// Addition and assignment operator.
-    friend LatticeSite operator+=(const LatticeSite &latticeSite, const Eigen::Vector3d &offset)
+    friend LatticeSite operator+=(const LatticeSite &latticeSite, const Eigen::Vector3i &offset)
     {
         LatticeSite latnbr = LatticeSite(latticeSite.index(), latticeSite.unitcellOffset() + offset);
         return latnbr;
@@ -107,7 +107,7 @@ private:
     size_t _index;
 
     /// Offset relative to the unit cell at the origin in units of lattice vectors.
-    Eigen::Vector3d _unitcellOffset;
+    Eigen::Vector3i _unitcellOffset;
 
 };
