+++ src/PyBinding.cpp
@@ -246,7 +246,7 @@ PYBIND11_MODULE(_icet, m)
 
         )pbdoc")
         .def(py::init<const int,
-                      const Vector3d &>(),
+                      const Vector3i &>(),
              "Initializes a LatticeSite object.",
              py::arg("site_index"),
              py::arg("unitcell_offset"))
@@ -262,7 +262,7 @@ PYBIND11_MODULE(_icet, m)
             "list(int) : unit cell offset (in units of the cell vectors)")
         .def(py::self < py::self)
         .def(py::self == py::self)
-        .def(py::self + Eigen::Vector3d())
+        .def(py::self + Eigen::Vector3i())
         .def("__hash__", [](const LatticeSite &latticeNeighbor)
              { return std::hash<LatticeSite>{}(latticeNeighbor); });
 
@@ -426,7 +426,7 @@ PYBIND11_MODULE(_icet, m)
                  return msg.str();
              })
         .def(py::self < py::self)
-        .def(py::self + Eigen::Vector3d())
+        .def(py::self + Eigen::Vector3i())
         .def(py::self += py::self);
 
     py::class_<OrbitList, std::shared_ptr<OrbitList>>(m, "_OrbitList",
